remove idea of "including" Common into connectors and consolidate all into the Base class and add the close_connection interface to base (which is required)
